pyproject.toml
[project]
name = "devmate"
version = "0.1.0"
description = "DevMate - AI-powered development assistant (interview project)"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "fastapi>=0.127.1",
    "langchain>=1.2.0",
    "langchain-openai>=0.1.0",
    "python-dotenv>=1.0.1",
    "uvicorn[standard]>=0.40.0",
    "tavily-python>=0.3.0",
    "mcp>=1.0.0",
    "colorlog>=6.10.1",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-community>=0.4.1",
    "dashscope>=1.25.5",
    "faiss-cpu>=1.13.2",
    "deepagents>=0.3.1",
]

[project.scripts]
devmate = "devmate.main:main"

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.0",
]
